Revert "Migrate openshift/origin/master to build01"
This reverts commit 915dfd13964c6782b6b6f622f11834a87e9394c7.
