+++ cmd/ci-operator-prowgen/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/sirupsen/logrus"
 	kubeapi "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
+	"k8s.io/apimachinery/pkg/util/sets"
 	v1 "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	prowconfig "k8s.io/test-infra/prow/config"
 
@@ -21,7 +22,6 @@ import (
 	"github.com/openshift/ci-tools/pkg/config"
 	"github.com/openshift/ci-tools/pkg/jobconfig"
 	jc "github.com/openshift/ci-tools/pkg/jobconfig"
-	"github.com/openshift/ci-tools/pkg/migrate"
 	"github.com/openshift/ci-tools/pkg/promotion"
 )
 
@@ -139,7 +139,7 @@ func generatePodSpec(info *prowgenInfo, secrets []*cioperatorapi.Secret) *kubeap
 		ReadOnly:  true,
 	}}
 
-	if migrate.Migrated(info.Org, info.Repo, info.Branch) {
+	if migrated(info.Org, info.Repo) {
 		volumeMounts = []kubeapi.VolumeMount{
 			{
 				Name:      sentryDsnMountName,
@@ -166,7 +166,7 @@ func generatePodSpec(info *prowgenInfo, secrets []*cioperatorapi.Secret) *kubeap
 		},
 	}}
 
-	if migrate.Migrated(info.Org, info.Repo, info.Branch) {
+	if migrated(info.Org, info.Repo) {
 		volumes = []kubeapi.Volume{
 			{
 				Name: sentryDsnMountName,
@@ -263,7 +263,7 @@ func generateCiOperatorPodSpec(info *prowgenInfo, secrets []*cioperatorapi.Secre
 		fmt.Sprintf("--branch=%s", info.Branch),
 	}, additionalArgs...)
 
-	if migrate.Migrated(info.Org, info.Repo, info.Branch) {
+	if migrated(info.Org, info.Repo) {
 		ret.Containers[0].Args = append([]string{
 			"--give-pr-author-access-to-namespace=true",
 			"--artifact-dir=$(ARTIFACTS)",
@@ -389,7 +389,7 @@ func generatePodSpecTemplate(info *prowgenInfo, release string, test *cioperator
 	// TODO add to all templates when they are migrated
 	// TODO expose boskos (behind an oauth proxy) so it can be used by build clusters
 	if needsLeaseServer {
-		if migrate.Migrated(info.Org, info.Repo, info.Branch) {
+		if migrated(info.Org, info.Repo) {
 			container.Args = append(container.Args, "--lease-server=https://boskos-ci.svc.ci.openshift.org")
 			container.Args = append(container.Args, "--lease-server-username=ci")
 			container.Args = append(container.Args, "--lease-server-password-file=/etc/boskos/password")
@@ -460,6 +460,14 @@ func generatePodSpecTemplate(info *prowgenInfo, release string, test *cioperator
 				Value: fmt.Sprintf("https://rpms.svc.ci.openshift.org/openshift-origin-v%s/", conf.PreviousVersion)})
 	}
 	return podSpec
+}
+
+var (
+	migratedRepos = sets.NewString("openshift/ci-secret-mirroring-controller")
+)
+
+func migrated(org, repo string) bool {
+	return migratedRepos.Has(fmt.Sprintf("%s/%s", org, repo))
 }
 
 func generatePodSpecRandom(info *prowgenInfo, test *cioperatorapi.TestStepConfiguration) *kubeapi.PodSpec {
@@ -659,12 +667,8 @@ func generateJobs(
 		}
 
 		if element.Cron == nil {
-			presubmit := *generatePresubmitForTest(element.As, info, label, podSpec, true, configSpec.CanonicalGoRepository)
-			presubmit.Cluster = migrate.GetBuildClusterForPresubmit(info.Org, info.Repo, info.Branch)
-			presubmits[orgrepo] = append(presubmits[orgrepo], presubmit)
+			presubmits[orgrepo] = append(presubmits[orgrepo], *generatePresubmitForTest(element.As, info, label, podSpec, true, configSpec.CanonicalGoRepository))
 		} else {
-			periodic := *generatePeriodicForTest(element.As, info, label, podSpec, true, *element.Cron, configSpec.CanonicalGoRepository)
-			periodic.Cluster = migrate.GetBuildClusterForPeriodic(info.Org, info.Repo, info.Branch)
 			periodics = append(periodics, *generatePeriodicForTest(element.As, info, label, podSpec, true, *element.Cron, configSpec.CanonicalGoRepository))
 		}
 	}
@@ -687,15 +691,11 @@ func generateJobs(
 			presubmitTargets = append(presubmitTargets, "[release:latest]")
 		}
 		podSpec := generateCiOperatorPodSpec(info, nil, presubmitTargets)
-		presubmit := *generatePresubmitForTest("images", info, label, podSpec, true, configSpec.CanonicalGoRepository)
-		presubmit.Cluster = migrate.GetBuildClusterForPresubmit(info.Org, info.Repo, info.Branch)
-		presubmits[orgrepo] = append(presubmits[orgrepo], presubmit)
+		presubmits[orgrepo] = append(presubmits[orgrepo], *generatePresubmitForTest("images", info, label, podSpec, true, configSpec.CanonicalGoRepository))
 
 		if configSpec.PromotionConfiguration != nil {
 			podSpec := generateCiOperatorPodSpec(info, nil, imageTargets, []string{"--promote"}...)
-			postsubmit := *generatePostsubmitForTest("images", info, label, podSpec, configSpec.CanonicalGoRepository)
-			postsubmit.Cluster = migrate.GetBuildClusterForPostsubmit(info.Org, info.Repo, info.Branch)
-			postsubmits[orgrepo] = append(postsubmits[orgrepo], postsubmit)
+			postsubmits[orgrepo] = append(postsubmits[orgrepo], *generatePostsubmitForTest("images", info, label, podSpec, configSpec.CanonicalGoRepository))
 		}
 	}
