+++ pkg/jobconfig/files.go
@@ -415,6 +415,7 @@ func mergePresubmits(old, new *prowconfig.Presubmit) prowconfig.Presubmit {
 	merged.Optional = old.Optional
 	merged.MaxConcurrency = old.MaxConcurrency
 	merged.SkipReport = old.SkipReport
+	merged.Cluster = old.Cluster
 
 	return merged
 }
@@ -426,6 +427,7 @@ func mergePostsubmits(old, new *prowconfig.Postsubmit) prowconfig.Postsubmit {
 	merged := *new
 
 	merged.MaxConcurrency = old.MaxConcurrency
+	merged.Cluster = old.Cluster
 
 	return merged
 }
@@ -437,6 +439,7 @@ func mergePeriodics(old, new *prowconfig.Periodic) prowconfig.Periodic {
 	merged := *new
 
 	merged.MaxConcurrency = old.MaxConcurrency
+	merged.Cluster = old.Cluster
 
 	return merged
 }
